1.6.2 Big Commit - Read the Extended Description
For 1.6.2, they made some changes to how textures and other resources
work. You will notice that "mods.learncraft.textures/*" no longer
exists, and it replaced by "assets.learncraft.textures/*" and
"assets.learncraft.sound". This assets folder needs to be moved to
"\forge\mcp\src\minecraft," (outside the learncraft directory) otherwise
the textures will not load. It is fine to have a copy of assets inside
the learncraft folder, but it won't do anything. The reason I included
the assets folder in the wrong location is because I added a new sound,
and I didn't want to do something weird like e-mail it to everyone.