1.6.2 Commit Continued - Extended Description
This should be the final tweak. I included the assets folder in the
wrong place simply because I added a few new sounds (to demonstrate that
we can easily add sounds now, which is pretty cool). The assets folder
should go in "forge\mcp\src\minecraft" instead of inside the
"learncraft" folder.